Rename narrowed reference builder files for deprecation
